// generated/src/app/api-models/claim-api-detail-inst-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimApiDetailInstW {
  claimNumber: string;
  lineNumber: string;
  detailSvcDate: string;
  detailThruDate: string;
  diagnosisPointer: string;
  placeServPointer: string;
  procedureCode: string;
  quantity: string;
  billedAmount: string;
  status: string;
  alternateProcCode: string;
  accommodationRate: string;
  nationalUnlabeled1: string;
  cobPatLiabCvrgAmt: string;
  fullCvrgAmt: string;
  ocAllowedAmt: string;
  adjudDate: string;
  procedureModifier: string;
}